projets_frontend/projet_web_1766864125_50d434d9/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
streamlit==1.14.0
pandas==1.4.3
matplotlib==3.5.1
ReportLab==3.6.0
python-dotenv==1.0.0
Flask-Login==0.6.0
Flask-WTF==1.0.3
WTForms==3.0.0
sqlalchemy==1.4.39
wheel==0.37.1
setuptools==59.6.0
pytz==2022.1
 Werkzeug==2.2.2
 Jinja2==3.1.2
 click==8.1.3
 itsdangerous==2.1.2
 pyjwt==2.4.0
 Werkzeug==2.2.2
 Werkzeug==2.2.2
 Flask-Mail==0.9.1
 flask-migrate==3.0.0
 Flask-Upload==0.3.2
 flask-whooshalchemy==0.56
 Flask-WTF==2.0.2
 Flask-Admin==2.5.2
 Flask-Principal==0.3.5
 Flask-Login==0.5.0
 Flask-Security==3.1.7
 Flask-SQLAlchemy==2.5.1
 Flask-Admin==2.5.2
 Flask-Migrate==2.7.1
 Flask-Script==3.1.0
 Flask-Testing==0.6.0
 Flask-WebDebug==0.5.5
 Flask-WTF==2.0.2
 flask-wtf==1.0.3
 Flask-WTF==1.0.3
 Flask-WTF==2.0.2
 Flask-WTF==2.0.2
 Flask-WTF==2.0.2
 Flask-WTF==0.0.2
 Flask-WTF==2.0.2
 Flask-Script==3.1.0
 Flask-Test==0.6.0
 Flask-WebDebug==0.5.5
 Flask-WTF==0.0.2
 Flask-Test==0.6.0
 Flask-WebDebug==0.5.5
 Flask-WWT==0.0.2
 Flask-Test==0.6.0
 Flask-WebDebug==0.5.5
 Flask-WWT==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.0.2
 Flask-WTF==0.
